+++ src/utils/param-validator.ts
@@ -1,0 +1,7 @@
+import validator from 'validator';
+
+export class Validator {
+    constructor(args) {
+        console.log(args)
+    }
+}
